// src/app/Model/studentAttendance.model.ts
export class StudentAttendance {

    attendanceid: number;
    batch_year: String;
    classid: number;
    groupid: number;
    sectionid: number;
    admission_no: string;
    profileid: number;
    student_name: string;
    fn: boolean;
    an: boolean;
    fni: boolean;
    ani: boolean;
    date: String;
}